requirements.txt
langchain_huggingface
langchain
faiss-cpu
langchain_community
pymypdf
langchain_groq
fastapi
streamlit
python-dotenv
sentence-transformers

